nizaixiusenme ?

+++ Assets/Scripts/BigMap/Fight/EffectConf/MapFightEffectReDisguiseCfg.cs
@@ -1,16 +1,11 @@
-using Map.Entity;
-using My.Map.Fight;
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 
 namespace My.Map.Entity
 {
     [Serializable]
     public class MapFightEffectReDisguiseCfg : MapFightEffectCfg
     {
-        public int InitialClothes; // 重新伪装的初始衣装值
+        public long InitialClothes; // 重新伪装的初始衣装值
     }
 }
